pyproject.toml
[project]
name = "mango"
version = "0.1.0"
description = "Mango — Adaptable Graph Network Simulators via Meta-Learning"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "lightning>=2.5.6",
    "plotly>=6.4.0",
    "torch~=2.8",
    "torch-geometric>=2.7.0",
    "matplotlib>=3.10.7",
    "wandb>=0.22.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mango*"]
